+++ tests/test_06_math/test_621_bspline.py
@@ -83,8 +83,11 @@ def test_bspline_insert_knot():
     assert len(bspline.control_points) == 9
 
 
-def test_basis_vector():
-    pass
+def test_transform_interface():
+    from ezdxf.math import Matrix44
+    spline = BSpline(control_points=[(1, 0, 0), (3, 3, 0), (6, 0, 1)], order=3)
+    spline.transform(Matrix44.translate(1, 2, 3))
+    assert spline.control_points[0] == (2, 2, 3)
 
 
 DBSPLINE = [