+++ NEWS.md
@@ -42,6 +42,7 @@ Version 0.13b2 - dev
 - NEW: `BSpline.from_fit_points()` returns a B-Spline from fit points
 - NEW: `BSpline.arc_approximation()` returns a B-Spline approximation for an arc defined by `ConstructionArc()`
 - NEW: `BSpline.ellipse_approximation()` returns a B-Spline approximation for an ellipse defined by `ConstructionEllipse()`
+- NEW: `BSpline.transform()` transform B-spline by transformation matrix inplace
 - CHANGE: `Hatch` full support for rotated patterns.
 - CHANGE: `Hatch.set_pattern_definition()` added argument `angle` for pattern rotation. 
 - CHANGE: `Hatch.path.add_arc` renamed argument `is_counter_clockwise` to `ccw`, type `bool` and `True` by default